Split up contact modal into two components
